doc: mark toctree :hidden:, remove search/modindex
Package doc toctrees need to be marked `:hidden:` or their ToC will be included
in the main opendylan.org ToC. That's not a problem, per se, but it also
means (apparently) that the title of the document is included in the
opendylan.org ToC, which I don't want; I want each package's ToC entry to be
only the name of the package, for consistency.

Also removing the `modindex`, which is Python specific, and `search` which is
not needed with the Furo theme.