Modernize to match RFC 4648
This was prompted by @housel noting that `base64-encode` only accepted a string as
input. That function has been updated to accept any sequence of bytes or byte characters.

Padding with "=" is now supported.

The standard and URL-safe character sets documented in RFC 4648 are supported.

There is still no support for line breaks, whitespace, or streaming.